moved udprx functionality away from main and into a lib
