--- nix_send_udp_test.go
@@ -1,71 +1,0 @@
-// +build linux
-
-// Copyright 2018 Otis Elevator Company. All rights reserved.
-// Use of this source code is govered by the MIT license which
-// can be found in the LICENSE file.
-
-// Author: Jeremy Mill: jeremy.mill@otis.com
-
-// Otis udp_rx software has been designed to utilize information
-// security technology described in the Category 5 – Part 2 of the
-// Commerce Control List, within Part 774 of the Export Administration
-// Regulations (“EAR”)(15 CFR 774).  However, the Otis udp_rx software
-// has been made publicly available in accordance with Part 742.15(b)
-// of the EAR and is therefore not subject to U.S. export regulations.
-// Before downloading this software, be aware that the country in which
-// you are located may have restrictions related to the import, download,
-// possession, use and/or reexport of encryption items.  It is your
-// responsibility to comply with any applicable laws and regulations
-// pertaining the import, download, possession, use and/or reexport of
-// encryption items.
-
-package main
-
-import (
-	"net"
-	"testing"
-
-	log "github.com/sirupsen/logrus"
-)
-
-func TestCreateUDPSocket(t *testing.T) {
-	err := CreateUDPSocket()
-	if err != nil {
-		log.Fatal("Failed to create a socket", err)
-	}
-}
-
-func TestSendUDP(t *testing.T) {
-	go handleUDPConn(t)
-	srcip := "192.168.1.100"
-	destip := "127.0.0.1"
-	var srcprt uint = 55553
-	var destprt uint = 55552
-	databuf := make([]byte, 11)
-	for i := 0; i < 11; i++ {
-		databuf[i] = (byte)(10 - i)
-	}
-	SendUDP(srcip, destip, srcprt, destprt, databuf, 0)
-}
-func handleUDPConn(t *testing.T) {
-	listenAddr := ":55552"
-	ServerAddr, _ := net.ResolveUDPAddr("udp", listenAddr)
-	ServerConn, _ := net.ListenUDP("udp", ServerAddr)
-	defer ServerConn.Close()
-	buf := make([]byte, 1024)
-	for {
-		_, src, _ := ServerConn.ReadFromUDP(buf)
-		if src.IP.String() != "192.168.1.100" {
-			t.Errorf("src IP invalid: %s", src.IP.String())
-		}
-		if src.Port != 55553 {
-			t.Errorf("src Port invalid %d", src.Port)
-		}
-		break
-	}
-	for i := 0; i < 11; i++ {
-		if buf[i] != (byte)(10-i) {
-			t.Errorf("Data invalid")
-		}
-	}
-}
